+++ src/main/java/prgrms/marco/be02marbox/domain/exception/custom/Message.java
@@ -1,4 +1,4 @@
-package prgrms.marco.be02marbox.domain.user.exception;
+package prgrms.marco.be02marbox.domain.exception.custom;
 
 public enum Message {
 	DUPLICATE_EMAIL_EXP_MSG("이미 존재하는 이메일 입니다."),
